moved coconet class from seperate file into main program to avoid runtime error
